backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai==1.3.7
pypdf2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.7.0
numpy>=1.26.0
pandas>=2.0.0
requests==2.31.0
lz4==4.3.2
python-dateutil==2.8.2
typing-extensions>=4.8.0
charset-normalizer==3.2.0
langchain==0.0.27
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4

